requirements.txt
absl-py==2.2.1
ag2==0.8.4
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
attrs==25.3.0
autocommand==2.2.2
backports.tarfile==1.2.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
deepdiff==6.7.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
email_validator==2.2.0
extend_autogen==0.1.0
fastapi==0.115.12
fastapi-cli==0.0.7
fix-busted-json==0.0.18
Flask==3.1.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.0.0
inflect==7.3.1
iniconfig==2.1.0
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.49
langchain-text-splitters==0.3.7
langsmith==0.3.19
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mesop==1.0.0
more-itertools==10.3.0
msgpack==1.1.0
multidict==6.2.0
mypy-extensions==1.0.0
numpy==2.2.4
ollama==0.4.7
openai==1.69.0
opencv-python==4.11.0.86
ordered-set==4.1.0
orjson==3.10.16
packaging==24.2
pillow==11.1.0
platformdirs==4.2.2
pluggy==1.5.0
propcache==0.3.1
protobuf==6.30.2
pyautogen==0.8.4
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
Pygments==2.19.1
pytest==8.3.5
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
replicate==1.0.4
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.1
setuptools==75.8.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
tenacity==9.0.0
termcolor==2.5.0
tiktoken==0.9.0
tomli==2.0.1
tqdm==4.67.1
typeguard==4.3.0
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.4
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.18.3
zipp==3.19.2
zstandard==0.23.0
